New package: wlrctl-0.2.1
The previous version had a hardcoded wrksrc variable. This fixes it. thanks to zdykstra for making me realize this.

The previous version had a hardcoded wrksrc variable. This fixes it. thanks to zdykstra for making me realize this.#

The previous version had a hardcoded wrksrc variable. This fixes it. thanks to zdykstra for making me realize this.